pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "mysql-connector-python>=9.5.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-httpx>=0.36.0",
    "ruff>=0.14.6",
]
